### 2023-03-06 #5216 @editorialbot
[PRE REVIEW]: Bluebonnet: Scaling solutions for production analysis from unconventional oil and gas wells
**Submitting author:** @frank1010111 (<a href="http://orcid.org/0000-0002-3402-5578">Frank Male</a>)
**Repository:** https://github.com/frank1010111/bluebonnet/
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
