### 2023-03-06 #5216 @kyleniemeyer
Hello @frank1010111, thanks for your interest in JOSS. This pre-review issue is where we will assign an editor to handle your submission and find reviewers. As we are getting started, any reviewer recommendations you can provide would be welcome.
