### 2023-03-06 4986c0c
:memo: add JOSS badge to readme
