### 2023-03-07 #5216 @matthewfeickert
:wave: @leouieda @andrelreis @SaumikDana @AnantHariharan1996 @seisman As you are all JOSS reviewers with domain expertise in Geophysics and Python would you be willing to help review this submission? I have another reviewer who has already expressed interest over email in helping with this review, but it would be great to have at least two of you help review if possible. Please let me know if you have any questions.
