### 2023-03-14 #5216 @matthewfeickert
:wave: @gassmoeller @aguspesce As you are JOSS reviewers with domain expertise in Geophysics and Python would you be willing to help review this submission? Please let me know if you have any questions.
