### 2023-04-01 #5255 @AnantHariharan1996
Some -extremely minor- typographical points I've noted while reading the paper and documentation.
For the manuscript:

(1) In Figure 4 (in the rendered pdf version of the manuscript), the y-axis of the figure is being cut off so the reader can't really read "Pressure (psi)".

(2) The axis labels and in-text references to Figure 4 mention panels (a) and (b) but these labels are not present in Figure 4 itself (though they are obvious in context).

(3) Line 50: "The goal for producing this package is to provide researches..." -> Presumably, you mean "researchers" instead.

(4): Line 29: "...the user to estimate production for shale gas wells using a scaled solutions of the single-phase real gas diffusivity equation..." -> "...the user to estimate production for shale gas wells using scaled solutions of the single-phase real gas diffusivity equation..." (Remove the "a")

In the documentation, specifically: https://bluebonnet.readthedocs.io/en/latest/oil_flow.html

(1) In the third line, "characterestic" -> should be changed to "characteristic"

(2) In the background in the "Full three-phase flow" section, it looks like there is a rendering issue with the Walsh and Lake 03 differential equation, where the word "denominator" appears at the right end of the page and the actual equation is cut off.
