### 2023-04-01 frank1010111/bluebonnet#18 @AnantHariharan1996
[JOSS Review]: Issue loading Water Compressibility
I have been testing the fluids tool as documented in: https://bluebonnet.readthedocs.io/en/latest/fluids.html

I ran into an issue when trying to calculate water compressibility as in the following line:

from bluebonnet.fluids.water import compressibility_water_McCain

Which results in the following issue:

---->  from bluebonnet.fluids.water import compressibility_water_McCain

ImportError: cannot import name 'compressibility_water_McCain' from 'bluebonnet.fluids.water' (/Users/ananthariharan/opt/anaconda3/lib/python3.9/site-packages/bluebonnet/fluids/water.py).

Upon inspection of the version of water.py that I have downloaded, it looks like compressibility_water_McCain is not a module there, thought it appears to be in the most recent version on the repo in the src folder. Let me know if there's something different I should be doing or if I should be using a different version:

I installed the code using pip install bluebonnet.

This is part of the reviewing process for JOSS: https://github.com/openjournals/joss-reviews/issues/5255
