### 2023-04-02 frank1010111/bluebonnet#18 @frank1010111
I've released version 0.2.0 to PyPI, which should resolve this issue.
