### 2023-04-03 frank1010111/bluebonnet#19 @RyanSwindeman
[JOSS Review] KeyError when Well 20 data is missing
In the docs, specifically at [https://bluebonnet.readthedocs.io/en/latest/forecast_varying.html#fit-wells](https://bluebonnet.readthedocs.io/en/latest/forecast_varying.html#fit-wells), we use data from Well 20 that isn't directly included in the repo. Can you provide a note on using git lfs (or any other tool) to fetch that data? If left as is, we obtain the following error when running the example verbatim:
```
KeyError: "None of [Index(['Days', 'Gas', 'Pressure'], dtype='object')] are in the [columns]"
```

This is part of the reviewing process for JOSS: openjournals/joss-reviews#5255
