### 2023-04-03 frank1010111/bluebonnet#20 @RyanSwindeman
[JOSS Review] Documentation tweaks
Great job overall. Please consider the following slight adjustments:
- I’d add a note at the beginning saying these code snippets are meant to be run in a jupyter notebook. It'd be extra nice to include a jupyter notebook that demonstrates the same code in the documentation. It probably already exists since Magics are used there.
- Often there are trailing `;` that can be removed in the examples (often after `ax.set(...)`)
- In [Simulating Tight Oil Flow](https://bluebonnet.readthedocs.io/en/latest/oil_flow.html#single-phase-oil-simulation) the `( );`  encapsulating the `pvt_oil.plot` is unnecessary and can be removed
- In [Varying Fracface Pressure](https://bluebonnet.readthedocs.io/en/latest/forecast_varying.html#) I’m unclear of the meaning of `[Text(0.5, 0, 'Time'), None, Text(0, 0.5, 'Pressure')]` and suspect it should be removed

This is part of the reviewing process for JOSS: openjournals/joss-reviews#5255
