### 2023-04-03 frank1010111/bluebonnet#21 @RyanSwindeman
[JOSS Review] Environment details
2 questions associated with the Python environment:
- When installing bluebonnet, what versions of Python are supported? I see a closed PR for 3.11, but what's the lower bound? Can you include this answer in the documentation?
- Could you explicitly include pip dependencies somewhere (e.g. in a `requirements.txt` file)? I know they're listed in the paper acknowledgments.

This is part of the reviewing process for JOSS: https://github.com/openjournals/joss-reviews/issues/5255
