### 2023-04-03 frank1010111/bluebonnet#19 @frank1010111
Thank you for your comment. The data from Well 20 is available only after you install git-lfs. See [the contributing part of the documentation](https://bluebonnet.readthedocs.io/en/latest/contributing.html#get-started).

I chose to use git-lfs for this with the expectation that several large files might be added to the repository for illustrative purposes.

Hope this helps.
