### 2023-04-03 frank1010111/bluebonnet#21 @frank1010111
Thank you for your suggestions.

I will add python versions to the documentation in #22. For package dependencies, I'm reticent to include a `requirements.txt` for fear that it will drift out of agreement with the requirements specified in `pyproject.toml`
