### 2023-04-03 41135d5
:memo: add python version and dependencies to README
