### 2023-04-04 frank1010111/bluebonnet#20 @frank1010111
Thank you for your suggestions.

Yes, code snippets portion of the documentation is generated from [jupyter notebooks found here](https://github.com/frank1010111/bluebonnet/tree/main/docs).

I added the trailing semicolons to suppress extra outputs in the source jupyter notebooks.

I will remove the encapsulating parentheses in Simulating Tight Oil Flow and extra jupyter printout of `[Text(0.5, 0, 'Time'), None, Text(0, 0.5, 'Pressure')]` in Varying Fracface Pressure.
