### 2023-04-06 3ff2592
:fire: remove flake8 config and pylint nox session now that ruff has replaced them
