### 2023-04-18 frank1010111/bluebonnet#19 @RyanSwindeman
Woops. Confirmed. I have it working, but I should be more precise in wording. Can you please include a blurb in the readme (maybe just after the python env/packages) or in the documents (where this error would otherwise appear) explicitly stating `git lfs install`/`git lfs pull` needs to be run to grab large files? Right now, the only mention of git lfs I see is under `Contributing` in local development. This means users who grabbed the package from pip will see this error if running through the documentation in order.
