### 2023-04-18 frank1010111/bluebonnet#23 @RyanSwindeman
[JOSS REVIEW] NotImplementedError Remarks
@frank1010111 could you please comment on the `NotImplementedError`s we see in `flow/reservoir.py`'s `MultiPhaseReservoir`? Thanks for any information.

This is part of the reviewing process for JOSS: https://github.com/openjournals/joss-reviews/issues/5255
