### 2023-04-18 frank1010111/bluebonnet#19 @frank1010111
I decided the simplest solution was to move that particular file from git-lfs to git tracking. You can see the changes at #24.
