### 2023-04-18 frank1010111/bluebonnet#23 @frank1010111
Yes, this is planned future functionality in #3 that would be nice to have, but its effect on tight-oil production matching and forecasting has not yet been studied.
