### 2023-04-22 75e3c51
:recycle: allow IdealReservoir to be initialized without specifying a non-existent flowproperties instance
