### 2023-07-08 frank1010111/bluebonnet#26 @AnantHariharan1996
Loading Production Data in Documentation Examples
A minor question about the documentation for bluebonnet.flow:

In the documentation at: https://bluebonnet.readthedocs.io/en/latest/forecast_varying.html, in the "Fit wells" section, the production data is loaded using:
`prod_file = f"https://raw.githubusercontent.com/frank1010111/bluebonnet/main/data/dataset_1_well_{well_number}.csv"`
But the well data is loaded using:
`well_file = f"../data/WellData_{well_number}.csv"`
Would it be more consistent if the path to the production file also referred to a file in the local data folder, e.g. with

`prod_file = f"../data/dataset_1_well_{well_number}.csv"?
`

This is part of the reviewing process for JOSS: https://github.com/openjournals/joss-reviews/issues/5255
