### 2023-07-10 frank1010111/bluebonnet#26 @frank1010111
So that people can run the tutorial without cloning or downloading the repository, I plan to standardize it on the first format.
