### 2023-08-03 frank1010111/bluebonnet#28 @matthewfeickert
[JOSS REVIEW]: Paper revision notes
@frank1010111 Thanks for the nice work. I have some hopefully very easy to address comments/questions RE: the paper:

* This might be surprising to hear, but the paper is actually too long as is at 7 pages. For reference, please check out the ["What should my paper contain?" section of the Submitting a paper to JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and the [example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
   - How to shorten things up: My recommendation is to remove all the figures. These more belong in the docs or in a user guide, and Figure 4 would need to be redone anyway given the poor DPI it was saved with.
   - If you have additional questions on preparing the paper please feel free to ask.
* The "These tools are" preceding the enumeration

https://github.com/frank1010111/bluebonnet/blob/f4fda4afb748eb5147aab0b3d565f2f3bc78b68c/paper/paper.md?plain=1#L47-L53

is awkward as the grammatical style doesn't match (would be better to have colons separating the module name from the short description of its function). Consider revising this.
* The references look okay :+1:, however I just noted that there was no URL for your Ph.D. thesis. It is there in the `note` field in the `.bib` entry

https://github.com/frank1010111/bluebonnet/blob/f4fda4afb748eb5147aab0b3d565f2f3bc78b68c/paper/paper.bib#L2-L8

so consider promoting that from `note` to `url` as it would be nice to have that as an online reference.
