### 2023-08-03 #5255 @matthewfeickert
@frank1010111 Thanks for the nice work. I have some hopefully very easy to address comments/questions RE: the paper over on https://github.com/frank1010111/bluebonnet/issues/28.

(Also apologies for the weeks delay &mdash; I'm currently moving house.)
