### 2023-08-04 567eddc
:hammer: add external=true for nox paper session
