### 2023-08-04 92fc483
:memo: remove figures, fix flow in JOSS paper
