### 2023-08-09 #5255 @frank1010111
@matthewfeickert Thank you for your comments and shepherding the paper along. We have addressed all the concerns that you raised in https://github.com/frank1010111/bluebonnet/pull/29, which has now been merged.
