### 2023-08-10 e194a48
chore: Update pre-commit hooks
* Apply pre-commit hook updates that don't require code changes when
  applied.
   - Use https://github.com/psf/black-pre-commit-mirror which is 2x faster.
      + Update black from 22.3.0 -> 23.7.0.
   - Use https://github.com/astral-sh/ruff-pre-commit as new official repo.
   - Update codespell from v2.1.0 -> v2.2.5.
