### 2023-08-11 #5255 @frank1010111
Thank you @matthewfeickert

I've released version v0.2.1 and created a Zenodo archive of the release: https://doi.org/10.5281/zenodo.8240137.
