### 2023-08-14 1de4e9e
Create CITATION.cff
