+++ CITATION.cff
@@ -1,0 +1,44 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Male
+  given-names: Frank
+  orcid: "https://orcid.org/0000-0002-3402-5578"
+- family-names: Marder
+  given-names: Michael P.
+- family-names: Ruiz-Maraggi
+  given-names: Leopoldo M.
+- family-names: Lake
+  given-names: Larry W.
+contact:
+- family-names: Male
+  given-names: Frank
+  orcid: "https://orcid.org/0000-0002-3402-5578"
+doi: 10.5281/zenodo.8240137
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Male
+    given-names: Frank
+    orcid: "https://orcid.org/0000-0002-3402-5578"
+  - family-names: Marder
+    given-names: Michael P.
+  - family-names: Ruiz-Maraggi
+    given-names: Leopoldo M.
+  - family-names: Lake
+    given-names: Larry W.
+  date-published: 2023-08-14
+  doi: 10.21105/joss.05255
+  issn: 2475-9066
+  issue: 88
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5255
+  title: "Bluebonnet: Scaling solutions for production analysis from
+    unconventional oil and gas wells"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05255"
+  volume: 8
+title: "Bluebonnet: Scaling solutions for production analysis from
+  unconventional oil and gas wells"
